requirements.txt
pyyaml
pygments
jinja2
argparse
pytest
flake8
mypy